refactor(cactus-web): improve the *Field story styles
The styling here is much more centralized, so I didn't have to do much
to update the styling helpers. Aside from that, I just gave the stories
a max width so they don't look too weird.

CACTUS-799
