+++ modules/cactus-web/src/AccessibleField/AccessibleField.tsx
@@ -6,9 +6,8 @@ import { margin, MarginProps, width, WidthProps } from 'styled-system'
 import { FieldWrapper } from '../FieldWrapper/FieldWrapper'
 import { Flex } from '../Flex/Flex'
 import { isFocusOut } from '../helpers/events'
-import { flexItem, FlexItemProps } from '../helpers/flexItem'
 import { omitProps } from '../helpers/omit'
-import { styledUnpoly, styledWithClass } from '../helpers/styled'
+import { flexItem, FlexItemProps, styledUnpoly, styledWithClass } from '../helpers/styled'
 import useId from '../helpers/useId'
 import Label, { LabelProps } from '../Label/Label'
 import StatusMessage, { Status } from '../StatusMessage/StatusMessage'
