+++ modules/cactus-web/src/DateInputField/DateInputField.story.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Box, DateInputField } from '../'
+import { DateInputField } from '../'
 import { FIELD_ARGS, HIDE_CONTROL, Story, STRING } from '../helpers/storybook'
 
 export default {
@@ -14,12 +14,13 @@ export default {
     ...FIELD_ARGS,
   },
   args: { label: 'Date Input Field', name: 'date-field', type: 'date', disabled: false },
+  parameters: { cactus: { overrides: { maxWidth: '500px' } } },
 } as const
 
 export const DefaultUsage: Story<typeof DateInputField> = (args) => (
-  <Box width="350px">
+  <>
     <DateInputField {...args} />
 
     <DateInputField {...args} disabled label="Date Input Field Disabled" />
-  </Box>
+  </>
 )
