+++ modules/cactus-web/src/RadioButtonField/RadioButtonField.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import styled from 'styled-components'
 
 import FieldWrapper from '../FieldWrapper/FieldWrapper'
-import { FlexItemProps } from '../helpers/flexItem'
 import { extractFieldStyleProps } from '../helpers/omit'
+import { FlexItemProps } from '../helpers/styled'
 import useId from '../helpers/useId'
 import Label, { LabelProps } from '../Label/Label'
 import RadioButton, { RadioButtonProps } from '../RadioButton/RadioButton'
