+++ modules/cactus-web/src/RadioCard/RadioCard.story.tsx
@@ -6,6 +6,7 @@ import { Action, FIELD_ARGS, Story } from '../helpers/storybook'
 export default {
   title: 'RadioCard',
   component: RadioCard,
+  parameters: { cactus: { overrides: { maxWidth: '700px' } } },
 } as const
 
 export const BasicUsage = (): React.ReactElement => {
